+++ swagger/CHANGES.md
@@ -1,6 +1,9 @@
 Change history of swagger
 =
 
+2014-01-03
+- (fix) do not add primitive type to the Api models
+
 2013-11-27
 - (fix) make Swagger work for WebServices with root ("/" or "") paths